[refactor] Extracted reusable helpers
